WebSocketConnection.cs manages web socket connection to live data server
