+++ Api/packages.config
@@ -6,4 +6,5 @@
   <package id="MathNet.Numerics" version="3.7.0" targetFramework="net45" />
   <package id="Newtonsoft.Json" version="7.0.1" targetFramework="net45" />
   <package id="RestSharp" version="105.2.3" targetFramework="net45" />
+  <package id="WebSocketSharp" version="1.0.3-rc11" targetFramework="net45" />
 </packages>
